pyproject.toml
[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 79

[tool.poetry]
name = "xfuse"
version = "0.1.3"
description = "Deep spatial data fusion"
authors = ["Ludvig Bergenstråhle <ludvig.bergenstrahle@scilifelab.se>"]

[tool.poetry.dependencies]
click = "^7.1.2"
h5py = "^3.0.0"
imageio = "^2.9.0"
matplotlib = "^3.3.2"
numpy = "^1.19.4"
opencv-python = "^4.4.0"
pandas = "^1.1.4"
Pillow = "^8.0.1"
pyro-ppl = "^1.5.0"
python = "^3.8"
scikit-learn = "^0.23.2"
scipy = "^1.5.4"
tensorboard = "^2.3.0"
tifffile = "^2020.10.1"
tomlkit = "^0.7.0"
torch = "^1.7.0"
torchvision = "^0.8.1"
tqdm = "^4.51.0"
tabulate = "^0.8.7"

[tool.poetry.dev-dependencies]
mypy = "^0.790"
pre-commit = "^2.8.2"
pylint = "^2.6.0"
pytest = "^6.1.2"
pytest-console-scripts = "^1.0.0"
pytest-cov = "^2.10.1"
pytest-datadir = "^1.3.1"
pytest-mock = "^3.3.1"

[tool.poetry.scripts]
xfuse = "xfuse.__main__:cli"
